Implemented GJK-EPA for mesh collision detection
Changes in the physics engine to add collision detection between meshes.
We detect the collision with the GJK algorithm and calculated the
penetration, contact normal and position with the EPA algorithm. Fixed
other minor errors

Replaced some push_back with emplace_back

Fixed error in the shaders that doesn't let us have illumination