Modified the Texture and Image classes
Fixed errors in the Texture class.

Now we load first the images with the image loader and then create the
texture from the images (in the utils package).

Set the minimum OpenGL version to 3.3 and made other changes in the
window class adn shaders.

Cleaned the Renderer classes to adapt them to the new Texture way of
binding and the RenderFlags type.