Added Basic Mesh Collision Detection and updated the Loader classes
Added a basic concave collision detection with meshes and fixed minor
errors in the collision detection package

Fixed minor error in the rigidbodies with the inifnite mass constructor

Changed some vector for sets and the convention about operators

Updated the Readers and loaders with a new RawMesh class that allows us
to use meshes with any package and not only by the graphics one. Also
added the name to the terrains loaded.