Modified project structure and cmake files
Now the FazeEngine is in its own library and the game in an executable

Moved the headers to the include folder and added pimpl idiom to some
classes so the headers don't import other implementation files

Added a namespace to the Engine library.

Removed the glew and glfw includes from many header files.

Changed some enums for enum classes.

Changed all the cmake library handling for targets and updated the git
tags. Also now cmake is working in Windows.

Added gtest and started some tests